\name{Multiclasstesting-package}
\alias{Multiclasstesting-package}
\docType{package}
\title{
Performance of N-ary Classification Testing
}
\description{
The function in this package allows the computation of statistical scores for the evaluation of the performances of a n-ary classification test. In fact, it extends the concepts of Sensitivity (Se), Specificity (Sp), Positive Predictive Value (PPV) and Negative Predictive Value (NPV) -defined for binary tests- to any number of classes.
}
\details{
\tabular{ll}{
Package: \tab Multiclasstesting\cr
Type: \tab Package\cr
Version: \tab 1.2.0\cr
Date: \tab 2009-06-15\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{
Nardini, C. and Liu, Y-H.

Maintainer: Christine Nardini \email{christine@picb.ac.cn}
}
\references{
C. Nardini, H. Peng, L. Wang, L. Benini, M.D. Kuo, MM-Correction: Meta-analysis-Based Multiple Hypotheses Correction in Omic Studies?Springer CCIS, Vol 25, pp 242-255, 2008.
}

\keyword{ package }

